pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gat"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "crewai",
    "pyyaml",
]

[project.scripts]
gat = "gat.cli:main"
